ecl.geo moved to ecl.util.geo.
tied ecl.util.geometry.

...
