+++ python/ecl/grid/__init__.py
@@ -1,7 +1,7 @@
 
 
 import ecl.util
-import ecl.geo
+import ecl.util.geometry
 
 from .cell import Cell
 from .ecl_grid import EclGrid
