+++ python/ecl/grid/faults/fault.py
@@ -17,7 +17,7 @@
 import numpy as np
 
 from ecl.util import monkey_the_camel
-from ecl.geo import Polyline, CPolyline, GeometryTools
+from ecl.util.geometry import Polyline, CPolyline, GeometryTools
 
 from .fault_line import FaultLine
 from .fault_segments import FaultSegment, SegmentMap
