+++ python/ecl/grid/faults/fault_block.py
@@ -20,7 +20,7 @@ from cwrap import BaseCClass
 from ecl.util import monkey_the_camel
 from ecl.util import DoubleVector, IntVector
 from ecl import EclPrototype
-from ecl.geo import Polyline, GeometryTools, CPolylineCollection
+from ecl.util.geometry import Polyline, GeometryTools, CPolylineCollection
 
 class FaultBlockCell(object):
