+++ python/ecl/summary/__init__.py
@@ -1,5 +1,5 @@
 import ecl.util
-import ecl.geo
+import ecl.util.geometry
 
 from .ecl_sum_var_type import EclSumVarType
 from .ecl_sum_tstep import EclSumTStep
