+++ python/ecl/util/CMakeLists.txt
@@ -27,3 +27,4 @@ set(PYTHON_SOURCES
 add_python_package("python.ecl.util"  ${PYTHON_INSTALL_PREFIX}/ecl/util "${PYTHON_SOURCES}" True)
 
 add_subdirectory(enums)
+add_subdirectory(geometry)
