+++ python/ecl/util/geometry/CMakeLists.txt
@@ -10,4 +10,4 @@ set(PYTHON_SOURCES
     surface.py
 )
 
-add_python_package("python.ecl.geo"  ${PYTHON_INSTALL_PREFIX}/ecl/geo "${PYTHON_SOURCES}" True)
+add_python_package("python.ecl.util.geometry"  ${PYTHON_INSTALL_PREFIX}/ecl/util/geometry "${PYTHON_SOURCES}" True)
