+++ python/ecl/util/geometry/__init__.py
