+++ python/ecl/util/geometry/cpolyline.py
