+++ python/ecl/util/geometry/geometry_tools.py
