+++ python/ecl/util/geometry/polyline.py
