+++ python/ert/geo/__init__.py
@@ -1,8 +1,8 @@
-from ecl.geo import GeoPointset
-from ecl.geo import GeoRegion
-from ecl.geo import CPolyline
-from ecl.geo import CPolylineCollection
-from ecl.geo import Polyline
-from ecl.geo import XYZIo
-from ecl.geo import GeometryTools
-from ecl.geo import Surface
+from ecl.util.geometry import GeoPointset
+from ecl.util.geometry import GeoRegion
+from ecl.util.geometry import CPolyline
+from ecl.util.geometry import CPolylineCollection
+from ecl.util.geometry import Polyline
+from ecl.util.geometry import XYZIo
+from ecl.util.geometry import GeometryTools
+from ecl.util.geometry import Surface
