+++ python/tests/ecl_tests/test_fault_blocks.py
@@ -22,7 +22,7 @@ from ecl import EclDataType
 from ecl.eclfile import EclKW
 from ecl.grid import EclGrid, EclRegion
 from ecl.grid.faults import FaultBlock, FaultBlockLayer, FaultBlockCell,FaultCollection
-from ecl.geo import Polyline , CPolylineCollection
+from ecl.util.geometry import Polyline , CPolylineCollection
 from ecl.test import TestAreaContext
 from tests import EclTest
