+++ python/tests/ecl_tests/test_layer.py
@@ -20,7 +20,7 @@ import time
 
 from ecl.util import IntVector
 from ecl.grid import EclGrid
-from ecl.geo import CPolyline
+from ecl.util.geometry import CPolyline
 from ecl.grid.faults import Layer , FaultCollection
 from ecl.test import TestAreaContext
 from tests import EclTest
