+++ python/tests/geometry_tests/test_convex_hull.py
@@ -1,4 +1,4 @@
-from ecl.geo.geometry_tools import GeometryTools
+from ecl.util.geometry.geometry_tools import GeometryTools
 from tests import EclTest
