+++ python/tests/geometry_tests/test_cpolyline_collection.py
@@ -1,7 +1,7 @@
 import gc
 
-from ecl.geo import CPolylineCollection , CPolyline
-from ecl.geo.xyz_io import XYZIo
+from ecl.util.geometry import CPolylineCollection , CPolyline
+from ecl.util.geometry.xyz_io import XYZIo
 from ecl.test import TestAreaContext
 from tests import EclTest
 from ecl.util import DoubleVector
