+++ python/tests/geometry_tests/test_geo_pointset.py
@@ -1,4 +1,4 @@
-from ecl.geo import GeoPointset, Surface
+from ecl.util.geometry import GeoPointset, Surface
 from ecl.test import TestAreaContext
 from tests import EclTest
