+++ python/tests/geometry_tests/test_geo_region.py
@@ -1,4 +1,4 @@
-from ecl.geo import GeoRegion, GeoPointset, CPolyline, Surface
+from ecl.util.geometry import GeoRegion, GeoPointset, CPolyline, Surface
 from ecl.test import TestAreaContext
 from tests import EclTest
