+++ python/tests/geometry_tests/test_geometry_tools.py
@@ -1,7 +1,7 @@
 import math
 
-from ecl.geo import Polyline, GeometryTools , CPolyline
-from ecl.geo.xyz_io import XYZIo
+from ecl.util.geometry import Polyline, GeometryTools , CPolyline
+from ecl.util.geometry.xyz_io import XYZIo
 from ecl.test import TestAreaContext
 from tests import EclTest
