+++ python/tests/geometry_tests/test_intersection.py
@@ -1,4 +1,4 @@
-from ecl.geo import GeometryTools
+from ecl.util.geometry import GeometryTools
 from tests import EclTest
