+++ python/tests/geometry_tests/test_polygon_slicing.py
@@ -1,5 +1,5 @@
 from math import sqrt
-from ecl.geo.geometry_tools import GeometryTools
+from ecl.util.geometry.geometry_tools import GeometryTools
 from tests import EclTest
