+++ python/tests/geometry_tests/test_polyline.py
@@ -1,6 +1,6 @@
 
-from ecl.geo import Polyline, GeometryTools
-from ecl.geo.xyz_io import XYZIo
+from ecl.util.geometry import Polyline, GeometryTools
+from ecl.util.geometry.xyz_io import XYZIo
 from ecl.test import TestAreaContext
 from tests import EclTest
